add (int) casts to some strlen() invocations to prevent warnings

+++ src/buf.c
@@ -143,7 +143,7 @@ struct Buf *buf_strnappend (struct Buf *buf, const char *str, int n)
 /* Appends characters in str to buf. */
 struct Buf *buf_strappend (struct Buf *buf, const char *str)
 {
-	return buf_strnappend (buf, str, strlen (str));
+	return buf_strnappend (buf, str, (int) strlen (str));
 }
 
 /* appends "#define str def\n" */
